requirements.txt
tkinter==0.1.0
pandas==2.1.4
numpy==1.24.3
psycopg2-binary==2.9.9
requests==2.31.0
matplotlib==3.8.2
plotly==5.17.0
dash==2.16.1
dash-bootstrap-components==1.5.0
schedule==1.2.1
python-dateutil==2.8.2
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
tkcalendar==1.6.1
ttkthemes==3.2.2
akshare==1.12.85
tushare==1.2.89
yfinance==0.2.28
chardet==5.2.0